HSEARCH-2611 Delete timeout requirement for JSR352
In the module of WildFly integration tests, tests related to JSR-352 do
not require a specific transaction timeout setting. So, the timeout is
now reverted to previous value, defined by the classic MassIndexer
implementation.

The timeout issue raised previously in JSR-352 restart-ability test,
org.hibernate.search.test.integration.jsr352.massindexing.RestartIT.java,
was related to full-text query, and was not related to JSR-352.